Refactor image to move placholder to image
