--- packages/thumbprint-react/components/Image/components/picture.module.scss
@@ -1,8 +1,0 @@
-.imageLoading {
-    opacity: 0;
-}
-
-.imageLoaded {
-    opacity: 1;
-    transition: opacity 300ms;
-}
