+++ packages/thumbprint-react/components/Image/index.module.scss
@@ -1,0 +1,19 @@
+.image,
+.picture {
+    display: block;
+    width: 100%;
+}
+
+.image {
+    object-fit: cover;
+    object-position: center center;
+}
+
+.imageLoading {
+    opacity: 0;
+}
+
+.imageLoaded {
+    opacity: 1;
+    transition: opacity 300ms;
+}
